+++ ConsoleChess/Board/Enum/BoardSide.cs
@@ -1,0 +1,8 @@
+﻿namespace Board.Enum
+{
+    enum BoardSide
+    {
+        Top,
+        Bottom
+    }
+}