+++ ConsoleChess/Pieces/Piece.cs
@@ -1,4 +1,11 @@
-﻿using Pieces.Enum;
+﻿using Board;
+using Board.Enum;
+
+using Extensions;
+
+using Pieces.Enum;
+
+using System;
 
 namespace Pieces
 {
@@ -7,25 +14,85 @@ namespace Pieces
     /// </summary>
     abstract class Piece
     {
-        /// <summary>
-        /// Team of the piece.
-        /// </summary>
         /// <value>Gets the team of the piece.</value>
         public readonly Team Team;
-        
+
+        /// <value>Gets the board where the piece belongs.</value>
+        public readonly Board.Board Board;
+
         /// <summary>
-        /// Symbol of the piece on console.
+        /// <para>Contains the origin board side of the piece.</para>
+        /// This data is used to identify the direction to follow when the piece is of the ForwardMove type.
         /// </summary>
+        /// <value>Gets the origin side of the piece on the board.</value>
+        public readonly BoardSide Side;
+
+        /// <value>Gets the number of times that this piece has been moved.</value>
+        public int TimesMoved { get; private set; } = 0;
+        
         /// <value>Gets the symbol of the piece.</value>
         protected abstract string Symbol { get; }
+        /// <value>Contains the movement rule of the piece.</value>
+        protected abstract Movement Movement { get; }
+        
+        /// <value>
+        /// Gets the position of the piece in the <see cref="Board"/>.
+        /// </value>
+        /// <exception cref="NullReferenceException">Thrown when the piece is not in the <see cref="Board"/>.</exception>
+        public TwoDimensionPosition Position
+        {
+            get
+            {
+                for (int row = 0; row < Board.Rows; row++)
+                {
+                    for (int column = 0; column < Board.Columns; column++)
+                    {
+                        if (Board.GetPiece(row, column) == this)
+                        {
+                            return new TwoDimensionPosition(column, row);
+                        }
+                    }
+                }
+                throw new NullReferenceException("This piece is not in the Board.");
+            }
+        }
 
         /// <summary>
         /// Base contructor for a new generic piece.
         /// </summary>
         /// <param name="team">Team of the piece.</param>
-        protected Piece(Team team)
+        /// <param name="board">Board where the piece belongs.</param>
+        /// <param name="side">Origin side on the board.</param>
+        protected Piece(Team team, Board.Board board, BoardSide side)
         {
             Team = team;
+            Board = board;
+            Side = side;
+        }
+
+        /// <summary>
+        /// It must be called every time the piece is moved on the <see cref="Board"/>.
+        /// </summary>
+        public void Move()
+        {
+            TimesMoved++;
+        }
+
+        /// <summary>
+        /// Returns the possible targets for the piece.
+        /// </summary>
+        /// <returns>An array of booleans with the same dimensions as the <see cref="Board"/>, where true positions mean a possible target on the board.</returns>
+        public bool[,] PossibleTargets()
+        {
+            var movementSet = Movement.GetInvocationList();
+            var possibleTargets = new bool[Board.Rows, Board.Columns];
+
+            foreach (var movement in movementSet)
+            {
+                possibleTargets = possibleTargets.Merge(movement.DynamicInvoke(this) as bool[,]);
+            }
+
+            return possibleTargets;
         }
 
         /// <summary>